avoid zigzag on letters
There is a big problem when creating a map joining the characters: the letters are drawn at different y positions. To fix this, i place all chars at their original positions This makes the png a bit bigger, but at least it will appear correctly. Note: the ideal would be add a new parameter for this change.